Remove Long unboxing in OpalDefendantAccountService, use generic Mockito matcher for ParameterizedTypeReference in tests, and map Enforcer/Court IDs to int via safeInt per API (omit on overflow).

+++ src/main/java/uk/gov/hmcts/opal/dto/common/Enforcer.java
@@ -14,7 +14,7 @@ import uk.gov.hmcts.opal.dto.ToJsonString;
 public class Enforcer implements ToJsonString {
 
     @JsonProperty("enforcer_id")
-    private Long enforcerId;
+    private Integer enforcerId;
 
     @JsonProperty("enforcer_name")
     private String enforcerName;
